+++ src/postamble.js
@@ -129,6 +129,7 @@ function run(args = arguments_) {
 function run() {
 #endif
 
+#if !BOOTSTRAPPING_STRUCT_INFO
   if (runDependencies > 0) {
 #if RUNTIME_DEBUG
     dbg('run() called, but dependencies remain, so not running');
@@ -136,6 +137,7 @@ function run() {
     dependenciesFulfilled = run;
     return;
   }
+#endif
 
 #if PTHREADS || WASM_WORKERS
   if ({{{ ENVIRONMENT_IS_WORKER_THREAD() }}}) {
@@ -153,6 +155,7 @@ function run() {
 
   preRun();
 
+#if !BOOTSTRAPPING_STRUCT_INFO
   // a preRun added a dependency, run will be called later
   if (runDependencies > 0) {
 #if RUNTIME_DEBUG
@@ -161,6 +164,7 @@ function run() {
     dependenciesFulfilled = run;
     return;
   }
+#endif
 
   {{{ asyncIf(ASYNCIFY == 2) }}}function doRun() {
     // run may have just been called through dependencies being fulfilled just in this very frame,
@@ -292,6 +296,8 @@ function preInit() {
 #endif
 }
 
+var wasmExports;
+
 #if MODULARIZE == 'instance'
 // In MODULARIZE=instance mode we delay most of the initialization work until
 // the `init` function is called.
@@ -321,11 +327,6 @@ export default async function init(moduleArg = {}) {
   run();
 }
 
-#if (WASM_WORKERS || PTHREADS) && !WASM_ESM_INTEGRATION
-// When run as a worker thread run `init` immediately.
-if ({{{ ENVIRONMENT_IS_WORKER_THREAD() }}}) await init()
-#endif
-
 #if ENVIRONMENT_MAY_BE_NODE
 // When run as the main script under node we run `init` immediately.
 if (ENVIRONMENT_IS_NODE
@@ -347,10 +348,38 @@ if (ENVIRONMENT_IS_SHELL) {
 }
 #endif
 
-#else
-preInit();
-run();
+#else // MODULARIZE == instance
+
+#if WASM_WORKERS || PTHREADS
+if ({{{ ENVIRONMENT_IS_MAIN_THREAD() }}}) {
+// Call createWasm on startup if we are the main thread.
+// Worker threads call this once they receive the module via postMessage
 #endif
+
+#if WASM_ASYNC_COMPILATION
+
+#if MODULARIZE
+// In modularize mode the generated code is within a factory function so we
+// can use await here (since it's not top-level-await).
+wasmExports = await createWasm();
+#else
+// With async instantation wasmExports is assigned asynchronously when the
+// instance is received.
+createWasm();
+#endif
+
+#else
+wasmExports = createWasm();
+#endif
+
+#if WASM_WORKERS || PTHREADS
+}
+#endif
+
+preInit();
+{{{ runIfMainThread('run();') }}}
+
+#endif // MODULARIZE != instance
 
 #if BUILD_AS_WORKER
 
@@ -406,8 +435,4 @@ var workerResponded = false, workerCallbackId = -1;
   }
 })();
 
-#endif
-
-#if STANDALONE_WASM && ASSERTIONS && !WASM_BIGINT
-err('warning: running JS from STANDALONE_WASM without WASM_BIGINT will fail if a syscall with i64 is used (in standalone mode we cannot legalize syscalls)');
 #endif