+++ src/runtime_pthread.js
@@ -26,7 +26,7 @@ if (ENVIRONMENT_IS_PTHREAD) {
   // Thread-local guard variable for one-time init of the JS state
   var initializedJS = false;
 
-#if LOAD_SOURCE_MAP || USE_OFFSET_CONVERTER
+#if LOAD_SOURCE_MAP
   // When using postMessage to send an object, it is processed by the structured
   // clone algorithm.  The prototype, and hence methods, on that object is then
   // lost. This function adds back the lost prototype.  This does not work with
@@ -112,9 +112,6 @@ if (ENVIRONMENT_IS_PTHREAD) {
 #if LOAD_SOURCE_MAP
         wasmSourceMap = resetPrototype(WasmSourceMap, msgData.wasmSourceMap);
 #endif
-#if USE_OFFSET_CONVERTER
-        wasmOffsetConverter = resetPrototype(WasmOffsetConverter, msgData.wasmOffsetConverter);
-#endif
 
 #if !WASM_ESM_INTEGRATION
 #if MINIMAL_RUNTIME
@@ -122,7 +119,13 @@ if (ENVIRONMENT_IS_PTHREAD) {
         Module['wasm'] = msgData.wasmModule;
         loadModule();
 #else
-        wasmModuleReceived(msgData.wasmModule);
+        wasmModule = msgData.wasmModule;
+#if MODULARIZE == 'instance'
+        init();
+#else
+        createWasm();
+        run();
+#endif
 #endif // MINIMAL_RUNTIME
 #endif
       } else if (cmd === 'run') {